+++ include/core/SkWriter32.h
@@ -45,9 +45,9 @@ public:
     ~SkWriter32();
 
     // return the current offset (will always be a multiple of 4)
-    uint32_t bytesWritten() const { return fSize; }
+    size_t bytesWritten() const { return fSize; }
     // DEPRECATED: use bytesWritten instead  TODO(mtklein): clean up
-    uint32_t  size() const { return this->bytesWritten(); }
+    size_t size() const { return this->bytesWritten(); }
 
     // Returns true if we've written only into the storage passed into constructor or reset.
     // (You may be able to use this to avoid a call to flatten.)
@@ -268,9 +268,9 @@ private:
     Block*      fHead;
     Block*      fTail;
     size_t      fMinSize;
-    uint32_t    fSize;
+    size_t      fSize;
     // sum of bytes written in all blocks *before* fTail
-    uint32_t    fWrittenBeforeLastBlock;
+    size_t      fWrittenBeforeLastBlock;
 
     bool isHeadExternallyAllocated() const {
         return fHead == &fExternalBlock;
